[v1.x] Rearrange ONNX tests in Nightly CI (#20075)
* split test_onnxruntime.py

* pytest with mark

* add pytest.ini

* use integration tag

* remove test_onnxruntime.py

* add test gpt

* fix shape
